% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{arrange}
\alias{arrange}
\title{Define the arrangement of chromatograms}
\usage{
arrange(...)
}
\arguments{
\item{group_by}{The column to group by (in the samples metadata)}
}
\value{
A function that takes and returns a Peakly object
}
\description{
Define the arrangement of chromatograms
}
