% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-chromatograms.R
\name{create_chromatograms_from_compound_discoverer}
\alias{create_chromatograms_from_compound_discoverer}
\alias{create_chromatograms_from_compound_discoverer,lcmsPlotDataContainer,list-method}
\title{Creates an instance of class \code{lcmsPlotDataContainer} from a
Compound Discoverer results object.}
\usage{
create_chromatograms_from_compound_discoverer(obj, options)

\S4method{create_chromatograms_from_compound_discoverer}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer}.}

\item{options}{A \code{list} representing the plot object's options.}
}
\value{
An instance of class \code{lcmsPlotDataContainer} with chromatograms
of the specified Compound Discoverer compounds.
}
\description{
Creates an instance of class \code{lcmsPlotDataContainer} from a
Compound Discoverer results object.
}
\keyword{internal}
