% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-chromatograms.R
\name{create_chromatograms_from_feature_ids}
\alias{create_chromatograms_from_feature_ids}
\alias{create_chromatograms_from_feature_ids,lcmsPlotDataContainer,list-method}
\title{Create an instance of class \code{lcmsPlotDataContainer} from feature IDs}
\usage{
create_chromatograms_from_feature_ids(obj, options)

\S4method{create_chromatograms_from_feature_ids}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer}.}

\item{options}{A \code{list} representing the plot object's options.}
}
\value{
An instance of class \code{lcmsPlotDataContainer} with chromatograms
of the specified feature IDs.
}
\description{
The input features are specified in the \code{options} list
under \code{chromatograms$features}.
}
\keyword{internal}
