% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-chromatograms.R
\name{create_chromatograms_from_features}
\alias{create_chromatograms_from_features}
\alias{create_chromatograms_from_features,lcmsPlotDataContainer,list-method}
\title{Creates an instance of class \code{lcmsPlotDataContainer} from a features \code{matrix}
or \code{data.frame}}
\usage{
create_chromatograms_from_features(obj, options)

\S4method{create_chromatograms_from_features}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer}.}

\item{options}{A \code{list} representing the plot object's options.}
}
\value{
An instance of class \code{lcmsPlotDataContainer} with chromatograms
of the specified features.
}
\description{
The input features are specified in the \code{options} list
under \code{chromatograms$features}.
}
\keyword{internal}
