% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{create_chromatograms_from_raw}
\alias{create_chromatograms_from_raw}
\alias{create_chromatograms_from_raw,lcmsPlotDataContainer,matrix,character,numeric,numeric-method}
\title{Creates a lcmsPlotDataContainer from raw sample files}
\usage{
create_chromatograms_from_raw(obj, features, sample_ids, ppm, rt_tol)

\S4method{create_chromatograms_from_raw}{lcmsPlotDataContainer,matrix,character,numeric,numeric}(obj, features, sample_ids, ppm, rt_tol)
}
\arguments{
\item{features}{A matrix of entries with mz and rt values}

\item{sample_ids}{The sample IDs to select}

\item{ppm}{The ppm error for the chromatograms}

\item{rt_tol}{The RT tolerance for the chromatograms}

\item{raw_data}{A list of mzR objects}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates a lcmsPlotDataContainer from raw sample files
}
