% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-chromatograms.R
\name{create_full_rt_chromatograms}
\alias{create_full_rt_chromatograms}
\alias{create_full_rt_chromatograms,lcmsPlotDataContainer,list-method}
\title{Creates an instance of class \code{lcmsPlotDataContainer} from
base peak chromatograms (BPC) or total ion chromatograms (TIC)}
\usage{
create_full_rt_chromatograms(obj, options)

\S4method{create_full_rt_chromatograms}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer}.}

\item{options}{A \code{list} representing the plot object's options.}
}
\value{
An instance of class \code{lcmsPlotDataContainer} with BPC or TIC.
}
\description{
The type of chromatograms is defined in
\code{options$chromatograms$aggregation_fun}.
}
\keyword{internal}
