% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-purity.R
\name{create_purity_scores}
\alias{create_purity_scores}
\title{Populate the \code{purity_scores} slot from a \code{purityA} object}
\usage{
create_purity_scores(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer} whose \code{data_obj}
slot is a \code{purityA} object.}

\item{options}{A \code{list} of plot options.}
}
\value{
A modified \code{lcmsPlotDataContainer} with the \code{purity_scores} slot
populated.
}
\description{
Extracts per-scan precursor ion purity scores from \code{purityA@puritydf} and
stores them in the \code{purity_scores} slot of the data container.
}
\keyword{internal}
