% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-spectra.R
\name{create_spectra}
\alias{create_spectra}
\alias{create_spectra,lcmsPlotDataContainer,list-method}
\title{Create an instance of class \code{lcmsPlotDataContainer} from spectra}
\usage{
create_spectra(obj, options)

\S4method{create_spectra}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer}.}

\item{options}{A \code{list} representing the plot object's options.}
}
\value{
An \code{lcmsPlotDataContainer} object with the created spectra.
}
\description{
Create an instance of class \code{lcmsPlotDataContainer} from spectra
}
\keyword{internal}
