% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-spectra.R
\name{create_spectra_for_sample}
\alias{create_spectra_for_sample}
\title{Create spectra for a single sample}
\usage{
create_spectra_for_sample(
  raw_obj,
  detected_peaks,
  sample_metadata,
  options,
  rt_range = NULL
)
}
\arguments{
\item{raw_obj}{An instance of class \code{mzR}.}

\item{detected_peaks}{A \code{data.frame} containing the detected peaks to
consider. Only applicable for modes \code{"closest_apex"} and \code{"across_peak"}.}

\item{sample_metadata}{A \code{data.frame} containing the sample's metadata.}

\item{options}{A \code{list} representing the plot object's options.}

\item{rt_range}{A \code{numeric} value indicating the RT range to apply to the
detected peaks.}
}
\value{
A \code{data.frame} representing spectra with columns \code{mz}, \code{intensity},
and \code{rt}.
}
\description{
Create spectra for a single sample
}
