% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{default_options}
\alias{default_options}
\title{Get the default options}
\usage{
default_options()
}
\value{
A named \code{list} containing all default options.
}
\description{
\code{default_options()} returns a list of default settings used throughout the
package. These options control aspects such as sample metadata handling,
plotting, chromatogram and spectra display, and general visualization
parameters.
}
\keyword{internal}
