% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-interface.R
\name{.app_navrail}
\alias{.app_navrail}
\title{Build the left nav rail (file upload, samples, tab nav)}
\usage{
.app_navrail()
}
\value{
A \code{shiny::tags$aside}.
}
\description{
Build the left nav rail (file upload, samples, tab nav)
}
\keyword{internal}
