% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-interface.R
\name{.app_topbar}
\alias{.app_topbar}
\title{Build the sticky topbar of the dashboard}
\usage{
.app_topbar()
}
\value{
A \code{shiny::tags$header}.
}
\description{
Build the sticky topbar of the dashboard
}
\keyword{internal}
