% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-launcher.R
\name{.build_server}
\alias{.build_server}
\title{Internal server builder for the Shiny app}
\usage{
.build_server()
}
\value{
A function suitable for \code{shiny::shinyApp(server = ...)}.
}
\description{
Internal server builder for the Shiny app
}
\keyword{internal}
