% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source-compound-discoverer-node.R
\name{.cd_node_build_compounds_feature}
\alias{.cd_node_build_compounds_feature}
\title{Feature-level compound builder (per-file feature topology)}
\usage{
.cd_node_build_compounds_feature(tabs)
}
\description{
Feature-level compound builder (per-file feature topology)
}
\keyword{internal}
