% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-module-chromatogram.R
\name{.chromatogram_ui}
\alias{.chromatogram_ui}
\title{UI for the BPC / TIC tab}
\usage{
.chromatogram_ui(id)
}
\arguments{
\item{id}{The module namespace id.}
}
\value{
A \code{shiny::tagList}.
}
\description{
UI for the BPC / TIC tab
}
\keyword{internal}
