% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-module-chromatogram.R
\name{.eic_server}
\alias{.eic_server}
\title{Server for the EIC tab}
\usage{
.eic_server(id, lcms_obj, selected_samples, metadata_cols)
}
\arguments{
\item{id}{The module namespace id.}

\item{lcms_obj}{Reactive returning the base \code{lcmsPlotClass}.}

\item{selected_samples}{Reactive returning the selected sample IDs.}

\item{metadata_cols}{Reactive returning available metadata column names.}
}
\value{
\code{NULL}.
}
\description{
Server for the EIC tab
}
\keyword{internal}
