% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-module-chromatogram.R
\name{.eic_ui}
\alias{.eic_ui}
\title{UI for the EIC tab}
\usage{
.eic_ui(id)
}
\arguments{
\item{id}{The module namespace id.}
}
\value{
A \code{shiny::tagList}.
}
\description{
UI for the EIC tab
}
\keyword{internal}
