% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source-lipid-search.R
\name{.ls_escape}
\alias{.ls_escape}
\title{Escape a string for literal use inside a regular expression}
\usage{
.ls_escape(x)
}
\description{
Escape a string for literal use inside a regular expression
}
\keyword{internal}
