% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-module-peak-density.R
\name{.peak_density_server}
\alias{.peak_density_server}
\title{Server for the peak density tab}
\usage{
.peak_density_server(id, lcms_obj, metadata_cols)
}
\arguments{
\item{id}{The module namespace id.}

\item{lcms_obj}{Reactive returning the base \code{lcmsPlotClass}.}

\item{metadata_cols}{Reactive returning available metadata column names.}
}
\value{
\code{NULL}.
}
\description{
Server for the peak density tab
}
\keyword{internal}
