% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-module-peak-density.R
\name{.peak_density_ui}
\alias{.peak_density_ui}
\title{UI for the peak density tab}
\usage{
.peak_density_ui(id)
}
\arguments{
\item{id}{The module namespace id.}
}
\value{
A \code{shiny::tagList}.
}
\description{
UI for the peak density tab
}
\keyword{internal}
