% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-module-spectra.R
\name{.spectra_server}
\alias{.spectra_server}
\title{Server for the spectra tab}
\usage{
.spectra_server(id, lcms_obj, metadata_cols)
}
\arguments{
\item{id}{The module namespace id.}

\item{lcms_obj}{Reactive returning the base \code{lcmsPlotClass}.}

\item{metadata_cols}{Reactive returning available metadata column names.}
}
\value{
\code{NULL}.
}
\description{
Server for the spectra tab
}
\keyword{internal}
