% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-module-spectra.R
\name{.spectra_ui}
\alias{.spectra_ui}
\title{UI for the spectra tab}
\usage{
.spectra_ui(id)
}
\arguments{
\item{id}{The module namespace id.}
}
\value{
A \code{shiny::tagList}.
}
\description{
UI for the spectra tab
}
\keyword{internal}
