% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-module-tic.R
\name{.tic_ui}
\alias{.tic_ui}
\title{UI for the TIC summary tab}
\usage{
.tic_ui(id)
}
\arguments{
\item{id}{The module namespace id.}
}
\value{
A \code{shiny::tagList}.
}
\description{
UI for the TIC summary tab
}
\keyword{internal}
