% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source.R
\docType{class}
\name{ExternalDataSource-class}
\alias{ExternalDataSource-class}
\title{External data source wrapper}
\description{
An S4 class representing an external data source such as MZmine.
This class acts as an adapter for external data sources by converting
data to a common format (XCMS).
}
\section{Slots}{

\describe{
\item{\code{name}}{A \code{character} value indicating the name of the data source.}

\item{\code{metadata}}{A \code{data.frame} representing the sample metadata.}

\item{\code{peaks}}{A \code{data.frame} representing the exported peaks.}
}}

