% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{from_XCMSnExp}
\alias{from_XCMSnExp}
\alias{from_XCMSnExp,XCMSnExp,character-method}
\title{Create a Peakly object from an XCMSnExp object}
\usage{
from_XCMSnExp(xcmsnExp, sample_id_column)

\S4method{from_XCMSnExp}{XCMSnExp,character}(xcmsnExp, sample_id_column)
}
\arguments{
\item{xcmsnExp}{The XCMSnExp object}

\item{sample_id_column}{Which column should be used as the sample ID}
}
\value{
A Peakly object
}
\description{
Create a Peakly object from an XCMSnExp object
}
