% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_features}
\alias{get_features}
\title{Get the feature data (m/z and RT ranges) for a set of features}
\usage{
get_features(
  options,
  sample_metadata,
  grouped_peaks = NULL,
  full_rt_range = NULL
)
}
\arguments{
\item{options}{The plot object's options. This contains the input features.}

\item{sample_metadata}{The sample's metadata.}

\item{grouped_peaks}{The grouped peaks to use as input features.}

\item{full_rt_range}{The full RT range if an RT range is not given.}
}
\value{
A list of feature data (see get_feature_data).
}
\description{
Get the feature data (m/z and RT ranges) for a set of features
}
\keyword{internal}
