% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_grouping_variables}
\alias{get_grouping_variables}
\title{Get the grouping variables for a plot}
\usage{
get_grouping_variables(opts)
}
\arguments{
\item{opts}{A \code{list} of plot options.}
}
\value{
A \code{character} vector containing the names of the grouping variables
used for faceting or grid layout. May contain \code{NULL} values if no grouping
is specified.
}
\description{
\code{get_grouping_variables()} extracts the variables used to group data in a
plot based on the provided plot options. It first checks for facet
specifications, and if absent, falls back to grid row/column settings.
}
\keyword{internal}
