% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_mz_range}
\alias{get_mz_range}
\title{Compute an m/z range given a ppm tolerance}
\usage{
get_mz_range(mz, ppm = 5)
}
\arguments{
\item{mz}{A \code{numeric} value indicating the m/z value
to calculate the range for.}

\item{ppm}{A \code{numeric} value indicating the tolerance
in parts per million (ppm). Default is 5.}
}
\value{
A \code{numeric} vector with two values indicating the m/z range.
}
\description{
Compute an m/z range given a ppm tolerance
}
\keyword{internal}
