% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{grid}
\alias{grid}
\title{Define the arrangement of chromatograms}
\usage{
grid(rows, cols)
}
\arguments{
\item{rows}{...}

\item{cols}{...}
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define the arrangement of chromatograms
}
