% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{intensity_map}
\alias{intensity_map}
\title{Define a 2D intensity map}
\usage{
intensity_map(mz_range, rt_range, sample_ids = NULL, density = FALSE)
}
\arguments{
\item{mz_range}{The m/z range of the map}

\item{rt_range}{The RT range of the map}

\item{sample_ids}{The sample IDs to select}

\item{density}{Whether to show a density or a point-cloud plot}
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define a 2D intensity map
}
\examples{
raw_files <- dir(
  system.file("cdf", package = "faahKO"),
  full.names = TRUE,
  recursive = TRUE)[1]

p <- lcmsPlot(raw_files) +
  intensity_map(mz_range = c(200, 600), rt_range = c(4200, 4500), density = TRUE)
}
