% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\name{io_close_raw_data}
\alias{io_close_raw_data}
\title{Close open \code{MsRawReader} connections}
\usage{
io_close_raw_data(raw_data)
}
\arguments{
\item{raw_data}{A list of \code{MsRawReader} objects (as returned by
\code{io_get_raw_data()}).}
}
\value{
\code{NULL}
}
\description{
Close open \code{MsRawReader} connections
}
\keyword{internal}
