% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{is_cd_node_source}
\alias{is_cd_node_source}
\title{Check whether an object is a Compound Discoverer scripting-node data source}
\usage{
is_cd_node_source(obj)
}
\arguments{
\item{obj}{An object to test.}
}
\value{
A \code{logical} value indicating whether the object is a
\code{CompoundDiscovererNodeSource}.
}
\description{
Check whether an object is a Compound Discoverer scripting-node data source
}
\keyword{internal}
