% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{is_cd_result}
\alias{is_cd_result}
\title{Check whether an object is a Compound Discoverer database connection}
\usage{
is_cd_result(obj)
}
\arguments{
\item{obj}{An object to test.}
}
\value{
A \code{logical} value indicating whether the object is a Compound
Discoverer database connection.
}
\description{
The object must inherit from \code{DBIConnection}, and its \code{dbname}
slot must reference a path ending in \code{.cdResult}.
}
\keyword{internal}
