% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{is_cd_results_path}
\alias{is_cd_results_path}
\title{Check whether a path corresponds to a Compound Discoverer results file}
\usage{
is_cd_results_path(path)
}
\arguments{
\item{path}{A \code{character} value giving a file system path.}
}
\value{
A \code{logical} value indicating whether the path corresponds to a
Compound Discoverer results directory.
}
\description{
Check whether a path corresponds to a Compound Discoverer results file
}
\keyword{internal}
