% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{is_lipid_search_source}
\alias{is_lipid_search_source}
\title{Check whether an object is a LipidSearch data source}
\usage{
is_lipid_search_source(obj)
}
\arguments{
\item{obj}{An object to test.}
}
\value{
A \code{logical} value indicating whether the object is a
\code{LipidSearchSource}.
}
\description{
Check whether an object is a LipidSearch data source
}
\keyword{internal}
