% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{is_mspurity_data}
\alias{is_mspurity_data}
\title{Check whether an object is an msPurity purityA result}
\usage{
is_mspurity_data(obj)
}
\arguments{
\item{obj}{The input object to check.}
}
\value{
A \code{logical} value.
}
\description{
Check whether an object is an msPurity purityA result
}
\keyword{internal}
