% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{is_xcms_object}
\alias{is_xcms_object}
\alias{is_xcms_object,lcmsPlotDataContainer-method}
\title{Check if the lcmsPlotDataContainer data object is an xcms object}
\usage{
is_xcms_object(object)

\S4method{is_xcms_object}{lcmsPlotDataContainer}(object)
}
\arguments{
\item{object}{The lcmsPlotDataContainer object}
}
\description{
Check if the lcmsPlotDataContainer data object is an xcms object
}
