% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{is_xcms_processed_data}
\alias{is_xcms_processed_data}
\title{Check whether an object is an XCMS data container}
\usage{
is_xcms_processed_data(obj)
}
\arguments{
\item{obj}{The input object to check.}
}
\value{
A \code{logical} value indicating whether \code{obj} is an
XCMS experiment object.
}
\description{
Check whether an object is an XCMS data container
}
\keyword{internal}
