% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{labels}
\alias{labels}
\title{Define the labels of the plot, such as title and legend}
\usage{
labels(title = NULL, legend = NULL)
}
\arguments{
\item{title}{The plot title}

\item{legend}{The legend's title}
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define the labels of the plot, such as title and legend
}
