% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{legend}
\alias{legend}
\title{Define the legend layout}
\usage{
legend(...)
}
\arguments{
\item{position}{The legend position}
}
\value{
A function that takes and returns a Peakly object
}
\description{
Define the legend layout
}
