% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{merge_by_index}
\alias{merge_by_index}
\title{Merge two data frames using a row-index match}
\usage{
merge_by_index(a, b, index_col)
}
\arguments{
\item{a}{The left data frame.}

\item{b}{The right data frame whose row order defines the index.}

\item{index_col}{The name of the column in \code{a} that contains row indices
referring to \code{b}.}
}
\value{
A \code{data.frame} resulting from a left join of \code{a} and the indexed \code{b}.
}
\description{
Merge two data frames using a row-index match
}
\keyword{internal}
