% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\name{ms_close}
\alias{ms_close}
\alias{ms_close,MzrReader-method}
\alias{ms_close,RawrrReader-method}
\title{Close a raw MS file reader connection}
\usage{
ms_close(reader)

\S4method{ms_close}{MzrReader}(reader)

\S4method{ms_close}{RawrrReader}(reader)
}
\arguments{
\item{reader}{An instance of \code{MsRawReader}.}
}
\value{
\code{invisible(NULL)}
}
\description{
Close a raw MS file reader connection
}
\keyword{internal}
