% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\docType{class}
\name{MzrReader-class}
\alias{MzrReader-class}
\title{An \code{mzR}-backed raw MS file reader}
\description{
Wraps an open \code{mzR} connection (as returned by \code{mzR::openMSfile()}).
}
\section{Slots}{

\describe{
\item{\code{connection}}{An \code{mzR} connection object.}
}}

\keyword{internal}
