% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound-discoverer.R
\name{open_cd_result_connection}
\alias{open_cd_result_connection}
\title{Open a connection to a Compound Discoverer results database}
\usage{
open_cd_result_connection(cd_result_path)
}
\arguments{
\item{cd_result_path}{A \code{character} value giving the path to a
Compound Discoverer results file.}
}
\value{
A \code{DBIConnection} object connected to the Compound Discoverer
SQLite database.
}
\description{
Establishes a read-only SQLite database connection to a Compound Discoverer
results directory.
}
\keyword{internal}
