% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\name{open_raw_reader}
\alias{open_raw_reader}
\title{Open a raw MS file as an \code{MsRawReader}}
\usage{
open_raw_reader(path)
}
\arguments{
\item{path}{A \code{character} value giving the file path.}
}
\value{
An instance of \code{MsRawReader} (either \code{MzrReader} or \code{RawrrReader}).
}
\description{
Dispatches to \code{MzrReader} for standard open formats (\code{.mzML}, \code{.mzXML},
\code{.CDF}) or \code{RawrrReader} for ThermoFisher \code{.raw} files.
}
\keyword{internal}
