% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\docType{class}
\name{Peakly-class}
\alias{Peakly-class}
\title{Peakly class}
\description{
Peakly class
}
\section{Slots}{

\describe{
\item{\code{samples}}{The samples metadata data frame}

\item{\code{features}}{The features (grouped raw peaks) data frame}

\item{\code{peaks}}{The raw peaks data frame}

\item{\code{options}}{The object options}

\item{\code{.data}}{The PeaklyChromatogramContainer object}

\item{\code{.plot}}{The underlying ggplot2 object}
}}

