% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly}
\alias{peakly}
\title{Create a Peakly object}
\usage{
peakly(dataset, sample_id_column = "sample_id")
}
\arguments{
\item{dataset}{An object of type XCMSnExp or MsExperiment}

\item{sample_id_column}{Which column should be used as the sample ID}
}
\value{
A Peakly object
}
\description{
Create a Peakly object
}
