% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly_arrange}
\alias{peakly_arrange}
\alias{peakly_arrange,Peakly,characterOrNULL-method}
\title{Set the arrangement of the chromatograms}
\usage{
peakly_arrange(peakly_obj, group_by)

\S4method{peakly_arrange}{Peakly,characterOrNULL}(peakly_obj, group_by)
}
\arguments{
\item{peakly_obj}{The Peakly object}

\item{group_by}{...}
}
\value{
The updated Peakly object
}
\description{
Set the arrangement of the chromatograms
}
