% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaklyChromatogramContainer-class.R
\name{peakly_chromatogram_container}
\alias{peakly_chromatogram_container}
\title{Create a PeaklyChromatogramContainer object}
\usage{
peakly_chromatogram_container(chromatograms_df, mass_traces_df, metadata_df)
}
\arguments{
\item{chromatograms_df}{A data frame containing columns intensity, rt, sample_id, and any additional metadata columns}
}
\value{
A PeaklyChromatogramContainer object
}
\description{
Create a PeaklyChromatogramContainer object
}
