% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly_chromatogram_from_features}
\alias{peakly_chromatogram_from_features}
\alias{peakly_chromatogram_from_features,Peakly,character,character,numeric,numeric,logical-method}
\title{Create chromatograms from features}
\usage{
peakly_chromatogram_from_features(
  peakly_obj,
  feature_ids,
  sample_ids,
  ppm,
  rt_tol,
  highlight_peaks
)

\S4method{peakly_chromatogram_from_features}{Peakly,character,character,numeric,numeric,logical}(
  peakly_obj,
  feature_ids,
  sample_ids,
  ppm,
  rt_tol,
  highlight_peaks
)
}
\arguments{
\item{peakly_obj}{The Peakly object}

\item{feature_ids}{A vector of feature IDs to plot}

\item{sample_ids}{A vector of sample IDs to plot}

\item{ppm}{The ppm error for the chromatograms}

\item{rt_tol}{The RT tolerance for the chromatograms}

\item{highlight_peaks}{Whether to highlight the picked peaks}
}
\value{
The updated Peakly object
}
\description{
Create chromatograms from features
}
