% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly_chromatogram_from_raw}
\alias{peakly_chromatogram_from_raw}
\alias{peakly_chromatogram_from_raw,Peakly,matrix,character,numeric,numeric-method}
\title{Create chromatograms from features}
\usage{
peakly_chromatogram_from_raw(peakly_obj, features, sample_ids, ppm, rt_tol)

\S4method{peakly_chromatogram_from_raw}{Peakly,matrix,character,numeric,numeric}(peakly_obj, features, sample_ids, ppm, rt_tol)
}
\arguments{
\item{peakly_obj}{The Peakly object}

\item{samples}{...}

\item{rt_range}{...}

\item{mz_range}{...}
}
\value{
The updated Peakly object
}
\description{
Create chromatograms from features
}
