% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly_facets}
\alias{peakly_facets}
\alias{peakly_facets,Peakly,characterOrNULL,numericOrNULL,numericOrNULL-method}
\title{Set the arrangement of the chromatograms}
\usage{
peakly_facets(peakly_obj, facets, ncol, nrow)

\S4method{peakly_facets}{Peakly,characterOrNULL,numericOrNULL,numericOrNULL}(peakly_obj, facets, ncol, nrow)
}
\arguments{
\item{peakly_obj}{The Peakly object}

\item{facets}{...}

\item{ncol}{...}

\item{nrow}{...}
}
\value{
The updated Peakly object
}
\description{
Set the arrangement of the chromatograms
}
