% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly_grid}
\alias{peakly_grid}
\alias{peakly_grid,Peakly,characterOrNULL,characterOrNULL-method}
\title{Set the arrangement of the chromatograms}
\usage{
peakly_grid(peakly_obj, rows, cols)

\S4method{peakly_grid}{Peakly,characterOrNULL,characterOrNULL}(peakly_obj, rows, cols)
}
\arguments{
\item{peakly_obj}{The Peakly object}

\item{rows}{...}

\item{cols}{...}
}
\value{
The updated Peakly object
}
\description{
Set the arrangement of the chromatograms
}
