% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly_labels}
\alias{peakly_labels}
\alias{peakly_labels,Peakly,characterOrNULL,characterOrNULL-method}
\title{Set the labels of the plot}
\usage{
peakly_labels(peakly_obj, title, legend)

\S4method{peakly_labels}{Peakly,characterOrNULL,characterOrNULL}(peakly_obj, title, legend)
}
\arguments{
\item{peakly_obj}{The Peakly object}

\item{title}{The title of the plot}

\item{legend}{The legend's title}
}
\value{
The updated Peakly object
}
\description{
Set the labels of the plot
}
