% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly_legend}
\alias{peakly_legend}
\alias{peakly_legend,Peakly,characterOrNULL-method}
\title{Set the legend options}
\usage{
peakly_legend(peakly_obj, position)

\S4method{peakly_legend}{Peakly,characterOrNULL}(peakly_obj, position)
}
\arguments{
\item{peakly_obj}{The Peakly object}

\item{position}{The legend's position in the plot}
}
\value{
The updated Peakly object
}
\description{
Set the legend options
}
