% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{peakly_mass_traces}
\alias{peakly_mass_traces}
\alias{peakly_mass_traces,Peakly-method}
\title{Set mass traces}
\usage{
peakly_mass_traces(peakly_obj)

\S4method{peakly_mass_traces}{Peakly}(peakly_obj)
}
\description{
Set mass traces
}
