% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaklyChromatogramContainer-class.R
\docType{class}
\name{PeaklyChromatogramContainer-class}
\alias{PeaklyChromatogramContainer-class}
\title{PeaklyChromatogramContainer class}
\description{
PeaklyChromatogramContainer class
}
\section{Slots}{

\describe{
\item{\code{rts}}{The chromatogram retention time points}

\item{\code{mzs}}{The mass trace masses}

\item{\code{intensities}}{The chromatogram intensities}

\item{\code{sample_ids}}{The sample IDs}

\item{\code{metadata}}{The metadata for each chromatogram point}
}}

