% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data}
\alias{plot_data}
\title{Plot the specified LC-MS data}
\usage{
plot_data(datasets, obj)
}
\arguments{
\item{datasets}{A \code{list} of data frames, each representing a dataset to be
visualised.}

\item{obj}{The lcmsPlot object.}
}
\value{
A \code{patchwork} plot object representing the final plot.
}
\description{
Plot the specified LC-MS data
}
\keyword{internal}
