% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-intensity-map.R
\name{plot_intensity_map}
\alias{plot_intensity_map}
\title{Plot an LC-MS intensity map}
\usage{
plot_intensity_map(datasets, supporting_datasets, options, single = FALSE)
}
\arguments{
\item{datasets}{A named \code{list} of data frames containing the primary datasets
to plot. For a 2D intensity map the used key is \code{intensity_maps}.}

\item{supporting_datasets}{A \code{list} of supporting data frames, such as
detected peaks, used for highlighting features.}

\item{options}{A list of plot options, controlling units, faceting,
highlighting, and other visual parameters.}

\item{single}{A \code{logical} value that indicates whether it should treat
the plot as a single dataset variant, which can affect faceting
and layout behavior. Default is \code{FALSE}.}
}
\value{
A \code{ggplot} object representing the 2D intensity map plot.
}
\description{
\code{plot_intensity_map()} generates a ggplot2 from an intensity map which is a
point-cloud of m/z and RT points.
}
\keyword{internal}
