% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mass-trace.R
\name{plot_mass_trace}
\alias{plot_mass_trace}
\title{Plot a mass trace}
\usage{
plot_mass_trace(datasets, supporting_datasets, options, single = FALSE)
}
\arguments{
\item{datasets}{A named \code{list} of data frames containing the primary datasets
to plot. For a mass trace plot the used key is \code{mass_traces}.}

\item{supporting_datasets}{A \code{list} of supporting data frames, such as
detected peaks, used for highlighting features.}

\item{options}{A list of plot options, controlling units, faceting,
highlighting, and other visual parameters.}

\item{single}{A \code{logical} value that indicates whether it should treat
the plot as a single dataset variant, which can affect faceting
and layout behavior. Default is \code{FALSE}.}
}
\value{
A \code{ggplot} object representing the mass trace plot.
}
\description{
\code{plot_mass_trace()} generates a ggplot2 from a mass trace
}
\keyword{internal}
