% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-total-ion-current.R
\name{plot_total_ion_current}
\alias{plot_total_ion_current}
\title{Plot total ion current for one or more samples}
\usage{
plot_total_ion_current(datasets, supporting_datasets, options, single = FALSE)
}
\arguments{
\item{datasets}{A named \code{list} of data frames containing the primary datasets
to plot. For a TIC plot the used key is \code{total_ion_current}.}

\item{supporting_datasets}{A \code{list} of supporting data frames, such as
detected peaks, used for highlighting features.}

\item{options}{A list of plot options, controlling units, faceting,
highlighting, and other visual parameters.}

\item{single}{A \code{logical} value that indicates whether it should treat
the plot as a single dataset variant, which can affect faceting
and layout behavior. Default is \code{FALSE}.}
}
\value{
A \code{ggplot} object representing the RT difference plot.
}
\description{
\code{plot_total_ion_current()} generates a ggplot2 of the total ion current
of the samples within the dataset.
}
\keyword{internal}
