% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{+,Peakly,function-method}
\alias{+,Peakly,function-method}
\title{Apply a function to the Peakly object}
\usage{
\S4method{+}{Peakly,function}(e1, e2)
}
\arguments{
\item{e1}{A Peakly object}

\item{e2}{A function that takes a Peakly object and returns another}
}
\description{
Apply a function to the Peakly object
}
