% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{remove_null_elements}
\alias{remove_null_elements}
\title{Remove \code{NULL} elements from a list}
\usage{
remove_null_elements(lst)
}
\arguments{
\item{lst}{A \code{list} from which \code{NULL} entries should be removed.}
}
\value{
A \code{list} containing only the non-\code{NULL} elements of \code{lst}.
}
\description{
Remove \code{NULL} elements from a list
}
\keyword{internal}
