% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{rt_diff_plot}
\alias{rt_diff_plot}
\title{Define the RT difference plot between raw and adjusted datasets}
\usage{
rt_diff_plot()
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define the RT difference plot between raw and adjusted datasets
}
