% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{rt_line}
\alias{rt_line}
\title{Define an vertical line across a retention time value}
\usage{
rt_line(intercept, line_type = "dashed", color = "black")
}
\arguments{
\item{intercept}{The x-axis intercept}

\item{line_type}{The line type}

\item{color}{The line color}
}
\description{
Define an vertical line across a retention time value
}
