% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{run_matching_plot_variant}
\alias{run_matching_plot_variant}
\title{Selects and executes the appropriate plot variant}
\usage{
run_matching_plot_variant(datasets, obj)
}
\arguments{
\item{datasets}{A \code{list} of data frames, each representing a dataset to be
visualised.}

\item{obj}{An instance of class \code{lcmsPlotClass}.}
}
\value{
A \code{patchwork} plot object generated by the selected plot variant.
}
\description{
Determines which plotting variant is applicable for the given datasets and
plotting options, then executes the first matching variant. Plot variants
are evaluated in order, and the first whose condition evaluates to \code{TRUE}
is used to generate the plot.
}
\keyword{internal}
