% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source-compound-discoverer-node.R
\name{show,CompoundDiscovererNodeSource-method}
\alias{show,CompoundDiscovererNodeSource-method}
\title{Show a summary of a \code{CompoundDiscovererNodeSource} object}
\usage{
\S4method{show}{CompoundDiscovererNodeSource}(object)
}
\arguments{
\item{object}{An instance of class \code{CompoundDiscovererNodeSource}.}
}
\value{
Invisible \code{NULL}
}
\description{
Show a summary of a \code{CompoundDiscovererNodeSource} object
}
