% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source-lipid-search.R
\name{show,LipidSearchSource-method}
\alias{show,LipidSearchSource-method}
\title{Show a summary of a \code{LipidSearchSource} object}
\usage{
\S4method{show}{LipidSearchSource}(object)
}
\arguments{
\item{object}{An instance of class \code{LipidSearchSource}.}
}
\value{
Invisible \code{NULL}
}
\description{
Show a summary of a \code{LipidSearchSource} object
}
