% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{show,Peakly-method}
\alias{show,Peakly-method}
\title{Plot the Peakly object}
\usage{
\S4method{show}{Peakly}(object)
}
\arguments{
\item{object}{The Peakly object}
}
\description{
Plot the Peakly object
}
