% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{total_ion_current}
\alias{total_ion_current}
\title{Define the total ion currents.}
\usage{
total_ion_current(sample_ids = NULL, type = "boxplot")
}
\arguments{
\item{sample_ids}{The sample IDs to select.}

\item{type}{The type of plot; one of "boxplot", "violin", "jitter".}
}
\description{
Define the total ion currents.
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:5]

p <- lcmsPlot(faahko, sample_id_column = "sample_name") +
  total_ion_current(type = "violin") +
  arrange(group_by = "sample_id") +
  labels(title = "Total ion current", legend = "Sample ID")
}
