% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\name{xcmsraw_to_readers}
\alias{xcmsraw_to_readers}
\title{Convert an \code{XcmsRawList} to a named list of \code{XcmsRawReader}s}
\usage{
xcmsraw_to_readers(xcmsraw_list)
}
\arguments{
\item{xcmsraw_list}{An \code{XcmsRawList} object.}
}
\value{
A named \code{list} of \code{XcmsRawReader} objects.
}
\description{
Each element is keyed by the file path stored in \verb{@filepath} of the
corresponding \code{xcmsRaw} object.
}
\keyword{internal}
