% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms-raw-list.R
\docType{class}
\name{XcmsRawList-class}
\alias{XcmsRawList-class}
\title{A list of \code{xcmsRaw} objects representing multiple samples}
\description{
\code{XcmsRawList} is a container for one or more \code{xcmsRaw} objects, where each
element corresponds to a single sample. It is the recommended way to pass
in-memory raw LC-MS data to \code{lcmsPlot}.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{list} of \code{xcmsRaw} objects, one per sample.}
}}

