% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms-raw-list.R
\name{XcmsRawList}
\alias{XcmsRawList}
\title{Create an \code{XcmsRawList} object}
\usage{
XcmsRawList(...)
}
\arguments{
\item{...}{One or more \code{xcmsRaw} objects, or a single \code{list} of \code{xcmsRaw}
objects.}
}
\value{
An instance of \code{XcmsRawList}.
}
\description{
Wraps one or more \code{xcmsRaw} objects into an \code{XcmsRawList} container for use
as the \code{dataset} argument of \code{lcmsPlot}.
}
\examples{
paths <- dir(
    system.file("cdf", package = "faahKO"),
    full.names = TRUE,
    recursive = TRUE
)[c(1, 2)]
raw1 <- xcms::xcmsRaw(paths[1])
raw2 <- xcms::xcmsRaw(paths[2])
xl <- XcmsRawList(raw1, raw2)
}
