% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{arrange}
\alias{arrange}
\title{Define the arrangement of chromatograms}
\usage{
arrange(group_by)
}
\arguments{
\item{group_by}{The column to group by (in the samples metadata)}
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define the arrangement of chromatograms
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:5]

p <- lcmsPlot(raw_files) +
  chromatogram(aggregation_fun = "max") +
  arrange(group_by = "sample_id") +
  legend(position = "bottom") +
  labels(legend = "Sample")
}
