% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source.R
\name{convert_rt_to_seconds}
\alias{convert_rt_to_seconds}
\title{Convert retention times to seconds}
\usage{
convert_rt_to_seconds(peaks)
}
\arguments{
\item{peaks}{A \code{data.frame} containing \code{rt}, \code{rtmin}, and \code{rtmax}
columns in minutes.}
}
\value{
\code{peaks} with retention time columns converted to seconds.
}
\description{
Convert retention time columns from minutes to seconds.
}
\keyword{internal}
