% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-chromatograms.R
\name{create_bpc_tic}
\alias{create_bpc_tic}
\title{Create a base peak or total ion current chromatogram.}
\usage{
create_bpc_tic(raw_data, aggregation_fun, rt_adjusted = NULL)
}
\arguments{
\item{raw_data}{An mzR object.}

\item{aggregation_fun}{The aggregation function to create the chromatogram.
One of 'sum' or 'max'.}

\item{rt_adjusted}{The RT adjusted values to use instead of the raw ones.}
}
\value{
A list with one data frame containing the chromatograms.
}
\description{
Create a base peak or total ion current chromatogram.
}
