% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-chromatograms.R
\name{create_chromatogram}
\alias{create_chromatogram}
\title{Create an extracted ion chromatogram.}
\usage{
create_chromatogram(
  raw_data,
  mz_range,
  rt_range,
  ms_level = 1,
  fill_gaps = FALSE
)
}
\arguments{
\item{raw_data}{An mzR object.}

\item{mz_range}{The m/z range of the chromatogram.}

\item{rt_range}{The RT range of the chromatogram.}

\item{ms_level}{The MS level of the scans to consider.}

\item{fill_gaps}{Whether to fill gaps between scans with zeros.}
}
\value{
A list  with chromatograms and mass traces data frames
}
\description{
Create an extracted ion chromatogram.
}
