% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{create_chromatograms_from_feature_ids}
\alias{create_chromatograms_from_feature_ids}
\alias{create_chromatograms_from_feature_ids,lcmsPlotDataContainer,character,character,numeric,numeric-method}
\title{Creates a lcmsPlotDataContainer from feature IDs}
\usage{
create_chromatograms_from_feature_ids(
  obj,
  feature_ids,
  sample_ids,
  ppm,
  rt_tol
)

\S4method{create_chromatograms_from_feature_ids}{lcmsPlotDataContainer,character,character,numeric,numeric}(
  obj,
  feature_ids,
  sample_ids,
  ppm,
  rt_tol
)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{feature_ids}{The feature IDs}

\item{sample_ids}{The sample IDs}

\item{ppm}{The ppm error for the chromatograms}

\item{rt_tol}{The RT tolerance for the chromatograms}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates a lcmsPlotDataContainer from feature IDs
}
