% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-chromatograms.R
\name{create_chromatograms_from_feature_ids}
\alias{create_chromatograms_from_feature_ids}
\alias{create_chromatograms_from_feature_ids,lcmsPlotDataContainer,list-method}
\title{Creates a lcmsPlotDataContainer from feature IDs}
\usage{
create_chromatograms_from_feature_ids(obj, options)

\S4method{create_chromatograms_from_feature_ids}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{options}{The options}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates a lcmsPlotDataContainer from feature IDs
}
