% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaklyChromatogramContainer-class.R
\name{create_chromatograms_from_features}
\alias{create_chromatograms_from_features}
\alias{create_chromatograms_from_features,list,data.frame,data.frame,data.frame,numeric,numeric-method}
\title{Creates a PeaklyChromatogramContainer from a features matrix}
\usage{
create_chromatograms_from_features(
  raw_data,
  features_matrix,
  peaks_matrix,
  samples,
  ppm,
  rt_tol
)

\S4method{create_chromatograms_from_features}{list,data.frame,data.frame,data.frame,numeric,numeric}(
  raw_data,
  features_matrix,
  peaks_matrix,
  samples,
  ppm,
  rt_tol
)
}
\arguments{
\item{raw_data}{A list of mzR objects}

\item{features_matrix}{The features matrix}

\item{peaks_matrix}{The peaks matrix}

\item{samples}{The samples metadata}

\item{ppm}{The ppm error for the chromatograms}

\item{rt_tol}{The RT tolerance for the chromatograms}
}
\value{
A PeaklyChromatogramContainer object
}
\description{
Creates a PeaklyChromatogramContainer from a features matrix
}
