% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{create_chromatograms_from_features}
\alias{create_chromatograms_from_features}
\alias{create_chromatograms_from_features,lcmsPlotDataContainer,list-method}
\title{Creates a lcmsPlotDataContainer from a features matrix}
\usage{
create_chromatograms_from_features(obj, options)

\S4method{create_chromatograms_from_features}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{options}{The options}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates a lcmsPlotDataContainer from a features matrix
}
