% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaklyChromatogramContainer-class.R
\name{create_chromatograms_from_raw}
\alias{create_chromatograms_from_raw}
\alias{create_chromatograms_from_raw,list,matrix,data.frame,numeric,numeric-method}
\title{Creates a PeaklyChromatogramContainer from raw sample files}
\usage{
create_chromatograms_from_raw(raw_data, features, samples, ppm, rt_tol)

\S4method{create_chromatograms_from_raw}{list,matrix,data.frame,numeric,numeric}(raw_data, features, samples, ppm, rt_tol)
}
\arguments{
\item{raw_data}{A list of mzR objects}

\item{features}{A matrix of entries with mz and rt values}

\item{samples}{The samples metadata}

\item{ppm}{The ppm error for the chromatograms}

\item{rt_tol}{The RT tolerance for the chromatograms}
}
\value{
A PeaklyChromatogramContainer object
}
\description{
Creates a PeaklyChromatogramContainer from raw sample files
}
