% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{create_data_container_from_obj}
\alias{create_data_container_from_obj}
\title{Create an lcmsPlotDataContainer object from a data object (e.g., XCMSnExp)}
\usage{
create_data_container_from_obj(data_obj, sample_id_column)
}
\arguments{
\item{data_obj}{...}

\item{sample_id_column}{...}
}
\description{
Create an lcmsPlotDataContainer object from a data object (e.g., XCMSnExp)
}
