% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-chromatograms.R
\name{create_full_rt_chromatograms}
\alias{create_full_rt_chromatograms}
\alias{create_full_rt_chromatograms,lcmsPlotDataContainer,list-method}
\title{Creates a lcmsPlotDataContainer from BPC or TIC data}
\usage{
create_full_rt_chromatograms(obj, options)

\S4method{create_full_rt_chromatograms}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{options}{The options}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates a lcmsPlotDataContainer from BPC or TIC data
}
