% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{create_full_rt_chromatograms}
\alias{create_full_rt_chromatograms}
\alias{create_full_rt_chromatograms,lcmsPlotDataContainer,character,character-method}
\title{Creates a lcmsPlotDataContainer from BPC or TIC data}
\usage{
create_full_rt_chromatograms(obj, sample_ids, aggregation_fun)

\S4method{create_full_rt_chromatograms}{lcmsPlotDataContainer,character,character}(obj, sample_ids, aggregation_fun)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{sample_ids}{The sample IDs to select}

\item{aggregation_fun}{Specify the function to be used to aggregate intensity
values across the mz value range for the same retention time.
Allowed values are "sum" (the default), "max"}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates a lcmsPlotDataContainer from BPC or TIC data
}
