% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-chromatograms.R
\name{create_full_rt_chromatograms}
\alias{create_full_rt_chromatograms}
\alias{create_full_rt_chromatograms,lcmsPlotDataContainer,list-method}
\title{Creates a lcmsPlotDataContainer from BPC or TIC data}
\usage{
create_full_rt_chromatograms(obj, options)

\S4method{create_full_rt_chromatograms}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{options}{The options}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates a lcmsPlotDataContainer from BPC or TIC data
}
\examples{
data_obj <- get_XCMSnExp_object_example(indices = c(1, 7))
data_container <- create_data_container_from_obj(data_obj, sample_id_column = "sample_name", metadata = NULL)

opts <- default_options()
opts$chromatograms$aggregation_fun <- 'max'
opts$chromatograms$sample_ids <- c('ko15', 'wt15')

data_container <- create_full_rt_chromatograms(data_container, opts)
}
