% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-intensity-map.R
\name{create_intensity_map}
\alias{create_intensity_map}
\alias{create_intensity_map,lcmsPlotDataContainer,list-method}
\title{Create an instance of class \code{lcmsPlotDataContainer} from an intensity map}
\usage{
create_intensity_map(obj, options)

\S4method{create_intensity_map}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer}.}

\item{options}{A \code{list} representing the plot object's options.}
}
\value{
An instance of class \code{lcmsPlotDataContainer} with the created
intensity map, a \code{data.frame} with columns \code{mz} and \code{rt}.
}
\description{
This function creates an \code{lcmsPlotDataContainer} object from
an intensity map which is a point-cloud of m/z and RT points.
}
\keyword{internal}
