% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{create_intensity_map}
\alias{create_intensity_map}
\alias{create_intensity_map,lcmsPlotDataContainer,list-method}
\title{Creates an lcmsPlotDataContainer from an intensity map}
\usage{
create_intensity_map(obj, options)

\S4method{create_intensity_map}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{options}{The lcmsPlot options}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates an lcmsPlotDataContainer from an intensity map
}
