% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-rt-diff.R
\name{create_rt_diff}
\alias{create_rt_diff}
\alias{create_rt_diff,lcmsPlotDataContainer,list-method}
\title{Create an instance of class \code{lcmsPlotDataContainer} from an RT adjustment
dataset}
\usage{
create_rt_diff(obj, options)

\S4method{create_rt_diff}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer}.}

\item{options}{A \code{list} representing the plot object's options.}
}
\value{
An instance of class \code{lcmsPlotDataContainer} with the created
RT adjustment dataset, a \code{tibble}
with columns \code{rt_raw}, \code{rt_adj}, and \code{diff}.
}
\description{
This function creates an \code{lcmsPlotDataContainer} object from
a dataset that contains the data to generate RT raw vs adjusted plots.
}
\keyword{internal}
