% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-rt-diff.R
\name{create_rt_diff}
\alias{create_rt_diff}
\alias{create_rt_diff,lcmsPlotDataContainer,list-method}
\title{Creates an lcmsPlotDataContainer from an RT diff dataset}
\usage{
create_rt_diff(obj, options)

\S4method{create_rt_diff}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{options}{The lcmsPlot options}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates an lcmsPlotDataContainer from an RT diff dataset
}
