% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-spectra.R
\name{create_spectra}
\alias{create_spectra}
\alias{create_spectra,lcmsPlotDataContainer,list-method}
\title{Creates/updates an lcmsPlotDataContainer from spectra}
\usage{
create_spectra(obj, options)

\S4method{create_spectra}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{options}{The lcmsPlot options}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates/updates an lcmsPlotDataContainer from spectra
}
