% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-spectra.R
\name{create_spectra_for_sample}
\alias{create_spectra_for_sample}
\title{Create spectra for a single sample.}
\usage{
create_spectra_for_sample(
  raw_obj,
  detected_peaks,
  sample_metadata,
  options,
  rt_range = NULL
)
}
\arguments{
\item{raw_obj}{An mzR object.}

\item{detected_peaks}{The detected peaks to consider.
Only for modes 'closest_apex' and 'across_peak'.}

\item{sample_metadata}{The sample's metadata.}

\item{options}{The plot object's options.}

\item{rt_range}{The RT range to apply to the detected peaks.}
}
\value{
A data frame representing spectra
with columns 'mz', 'intensity', and 'rt'.
}
\description{
Create spectra for a single sample.
}
