% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-spectra.R
\name{create_spectrum_from_closest_scan_to_rt}
\alias{create_spectrum_from_closest_scan_to_rt}
\title{Create a spectrum of the closest scan to the specified RT.}
\usage{
create_spectrum_from_closest_scan_to_rt(raw_data, rt, ms_level)
}
\arguments{
\item{raw_data}{An mzR object.}

\item{rt}{The RT to consider.}

\item{ms_level}{The MS level of the scans.}
}
\value{
A spectrum data frame.
}
\description{
Create a spectrum of the closest scan to the specified RT.
}
