% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-spectra.R
\name{create_spectrum_from_scan_index}
\alias{create_spectrum_from_scan_index}
\title{Create a spectrum of the specified scan.}
\usage{
create_spectrum_from_scan_index(raw_data, sample_metadata, scan_index)
}
\arguments{
\item{raw_data}{An mzR object.}

\item{sample_metadata}{The sample metadata.}

\item{scan_index}{The scan index.}
}
\description{
Create a spectrum of the specified scan.
}
