% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-total-ion-current.R
\name{create_total_ion_current}
\alias{create_total_ion_current}
\alias{create_total_ion_current,lcmsPlotDataContainer,list-method}
\title{Create an instance of class \code{lcmsPlotDataContainer} from total ion current
(TIC) dataset}
\usage{
create_total_ion_current(obj, options)

\S4method{create_total_ion_current}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{An instance of class \code{lcmsPlotDataContainer}.}

\item{options}{A \code{list} representing the plot object's options.}
}
\value{
An instance of class \code{lcmsPlotDataContainer} with the created
RT adjustment dataset, a \code{data.frame}
with columns \code{sample_id} and \code{intensity}.
}
\description{
This function creates an \code{lcmsPlotDataContainer} object from
a dataset that contains TIC values for each sample.
}
\keyword{internal}
