% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators-total-ion-current.R
\name{create_total_ion_current}
\alias{create_total_ion_current}
\alias{create_total_ion_current,lcmsPlotDataContainer,list-method}
\title{Creates an lcmsPlotDataContainer from a TIC}
\usage{
create_total_ion_current(obj, options)

\S4method{create_total_ion_current}{lcmsPlotDataContainer,list}(obj, options)
}
\arguments{
\item{obj}{A lcmsPlotDataContainer object}

\item{options}{The lcmsPlot options}
}
\value{
A lcmsPlotDataContainer object
}
\description{
Creates an lcmsPlotDataContainer from a TIC
}
