% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{default_options}
\alias{default_options}
\title{Get the default options.}
\usage{
default_options()
}
\value{
The default options.
}
\description{
Get the default options.
}
\examples{
def_opts <- default_options()
}
