% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{facets}
\alias{facets}
\title{Define plot's faceting}
\usage{
facets(facets, ncol = NULL, nrow = NULL)
}
\arguments{
\item{facets}{The facet factors from the sample metadata}

\item{ncol}{The number of columns}

\item{nrow}{The number of rows}
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define plot's faceting
}
