% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{facets}
\alias{facets}
\title{Define plot's faceting.}
\usage{
facets(facets, ncol = NULL, nrow = NULL, free_x = FALSE, free_y = FALSE)
}
\arguments{
\item{facets}{The facet factors from the sample metadata.}

\item{ncol}{The number of columns.}

\item{nrow}{The number of rows.}

\item{free_x}{Allow scales to vary across x.}

\item{free_y}{Allow scales to vary across y.}
}
\value{
A function that takes and returns a lcmsPlotClass object.
}
\description{
Define plot's faceting.
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:5]

p <- lcmsPlot(raw_files) +
  lcmsPlot::chromatogram(aggregation_fun = "max") +
  facets(facets = "sample_id")
}
