% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{facets}
\alias{facets}
\title{Define the arrangement of chromatograms}
\usage{
facets(facets, ncol = NULL, nrow = NULL)
}
\arguments{
\item{facets}{...}

\item{ncol}{...}

\item{nrow}{...}
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define the arrangement of chromatograms
}
