% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-adapters.R
\name{get_detected_peaks}
\alias{get_detected_peaks}
\alias{get_detected_peaks.character}
\alias{get_detected_peaks.XCMSnExp}
\title{Get the detected peaks from the data object (e.g. XCMSnExp)}
\usage{
get_detected_peaks(obj)

\method{get_detected_peaks}{character}(obj)

\method{get_detected_peaks}{XCMSnExp}(obj)
}
\arguments{
\item{obj}{The data object}
}
\description{
Get the detected peaks from the data object (e.g. XCMSnExp)
}
