% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-adapters.R
\name{get_detected_peaks}
\alias{get_detected_peaks}
\alias{get_detected_peaks.character}
\alias{get_detected_peaks.XCMSnExp}
\alias{get_detected_peaks.MsExperiment}
\title{Get the detected peaks from the data object (e.g. XCMSnExp)}
\usage{
get_detected_peaks(obj)

\method{get_detected_peaks}{character}(obj)

\method{get_detected_peaks}{XCMSnExp}(obj)

\method{get_detected_peaks}{MsExperiment}(obj)
}
\arguments{
\item{obj}{The data object}
}
\description{
Get the detected peaks from the data object (e.g. XCMSnExp)
}
\examples{
cdfs <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[c(1, 7)]
sample_names <- sub(basename(cdfs), pattern = ".CDF", replacement = "", fixed = TRUE)

pd <- data.frame(sample_name = sample_names,
                 sample_group = c("KO", "WT"),
                 stringsAsFactors = FALSE)

faahko <- readMsExperiment(spectraFiles = cdfs, sampleData = pd)

cwp <- xcms::CentWaveParam(peakwidth = c(20, 80), noise = 5000, prefilter = c(6, 5000))
faahko <- xcms::findChromPeaks(faahko, param = cwp)

detected_peaks <- get_detected_peaks(faahko)
}
