% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_feature_data}
\alias{get_feature_data}
\title{Get a feature's m/z and RT ranges given different feature specifications}
\usage{
get_feature_data(feature, options, full_rt_range)
}
\arguments{
\item{feature}{The input feature which can be a vector or a data frame row.
The accepted columns combinations are:
\itemize{
\item mz, rt (optional)
\item mzmin, mzmax, rtmin (optional), rtmax (optional)
}}

\item{options}{The plot object's options.}

\item{full_rt_range}{The full RT range if an RT range is not given.}
}
\value{
A named list defining a feature with names: feature_id, mzr, rtr.
}
\description{
Get a feature's m/z and RT ranges given different feature specifications
}
\keyword{internal}
