% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-adapters.R
\name{get_grouped_peaks}
\alias{get_grouped_peaks}
\alias{get_grouped_peaks.default}
\alias{get_grouped_peaks.XCMSnExp}
\alias{get_grouped_peaks.XcmsExperiment}
\alias{get_grouped_peaks.MsExperiment}
\title{Get the grouped peaks across samples (features) from the data object}
\usage{
get_grouped_peaks(obj)

\method{get_grouped_peaks}{default}(obj)

\method{get_grouped_peaks}{XCMSnExp}(obj)

\method{get_grouped_peaks}{XcmsExperiment}(obj)

\method{get_grouped_peaks}{MsExperiment}(obj)
}
\arguments{
\item{obj}{The data object}
}
\description{
Get the grouped peaks across samples (features) from the data object
}
\examples{
cdfs <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[c(1, 7)]
sample_names <- sub(basename(cdfs), pattern = ".CDF", replacement = "", fixed = TRUE)

pd <- data.frame(sample_name = sample_names,
                 sample_group = c("KO", "WT"),
                 stringsAsFactors = FALSE)

faahko <- MsExperiment::readMsExperiment(spectraFiles = cdfs, sampleData = pd)

cwp <- xcms::CentWaveParam(peakwidth = c(20, 80), noise = 5000, prefilter = c(6, 5000))
faahko <- xcms::findChromPeaks(faahko, param = cwp)

pdp <- xcms::PeakDensityParam(
    sampleGroups = pd$sample_group,
    minFraction = 0.4,
    bw = 30)
faahko <- xcms::groupChromPeaks(faahko, param = pdp)

grouped_peaks <- get_grouped_peaks(faahko)
}
