% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-adapters.R
\name{get_grouped_peaks}
\alias{get_grouped_peaks}
\alias{get_grouped_peaks.character}
\alias{get_grouped_peaks.XCMSnExp}
\alias{get_grouped_peaks.XcmsExperiment}
\alias{get_grouped_peaks.MsExperiment}
\title{Get the grouped peaks across samples (features) from the data object}
\usage{
get_grouped_peaks(obj)

\method{get_grouped_peaks}{character}(obj)

\method{get_grouped_peaks}{XCMSnExp}(obj)

\method{get_grouped_peaks}{XcmsExperiment}(obj)

\method{get_grouped_peaks}{MsExperiment}(obj)
}
\arguments{
\item{obj}{The data object}
}
\description{
Get the grouped peaks across samples (features) from the data object
}
