% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_grouping_variables}
\alias{get_grouping_variables}
\title{Get the grouping variables for a plot}
\usage{
get_grouping_variables(opts)
}
\arguments{
\item{opts}{The plot object's options.}
}
\value{
The grouping variables as a character vector.
}
\description{
Get the grouping variables for a plot
}
