% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-adapters.R
\name{get_metadata}
\alias{get_metadata}
\alias{get_metadata.character}
\alias{get_metadata.XCMSnExp}
\alias{get_metadata.MsExperiment}
\title{Get the metadata associated with the input object}
\usage{
get_metadata(obj, sample_id_column, metadata)

\method{get_metadata}{character}(obj, sample_id_column, metadata)

\method{get_metadata}{XCMSnExp}(obj, sample_id_column, metadata)

\method{get_metadata}{MsExperiment}(obj, sample_id_column, metadata)
}
\arguments{
\item{obj}{The data object}

\item{sample_id_column}{The column that should be associated with the sample ID}

\item{metadata}{The metadata data frame if not already included in the data object}
}
\value{
A data frame representing the metadata.
}
\description{
Get the metadata associated with the input object
}
\examples{
# Get metadata from sample paths
paths <- c("test01.mzML", "test02.mzML")
metadata <- get_metadata(paths, sample_id_column = NULL, metadata = NULL)
}
