% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{get_mz_range}
\alias{get_mz_range}
\title{Get the m/z range for an input m/z and a ppm error.}
\usage{
get_mz_range(mz, ppm = 5)
}
\arguments{
\item{mz}{The m/z value to calculate the range for.}

\item{ppm}{The ppm tolerance for the range.}
}
\value{
A vector with two values indicating the m/z range.
}
\description{
Get the m/z range for an input m/z and a ppm error.
}
