% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{get_plot}
\alias{get_plot}
\title{Get the underlying plot object}
\usage{
get_plot()
}
\description{
Get the underlying plot object
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:4]

p <- lcmsPlot(raw_files) +
  chromatogram(features = rbind(c(mzmin = 334.9, mzmax = 335.1, rtmin = 2700, rtmax = 2900))) +
  facets(facets = 'sample_id', ncol = 4) +
  rt_line(intercept = 2800, line_type = 'solid', color = 'red') +
  get_plot() +
  theme_bw()
}
