% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound-discoverer.R
\name{get_workflow_input_files}
\alias{get_workflow_input_files}
\title{Retrieve workflow input files from a Compound Discoverer database}
\usage{
get_workflow_input_files(conn)
}
\arguments{
\item{conn}{A \code{DBIConnection} to a Compound Discoverer results database.}
}
\value{
A \code{data.frame} containing information about the workflow input files.
}
\description{
Extracts the contents of the \code{WorkflowInputFiles} table from a
Compound Discoverer results database.
}
\keyword{internal}
