% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{grid}
\alias{grid}
\title{Define a gridded plot}
\usage{
grid(rows, cols)
}
\arguments{
\item{rows}{The factors that represent rows}

\item{cols}{The factors that represent columns}
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define a gridded plot
}
