% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{grid}
\alias{grid}
\title{Define the arrangement of chromatograms}
\usage{
grid(...)
}
\arguments{
\item{rows}{...}

\item{cols}{...}
}
\value{
A function that takes and returns a Peakly object
}
\description{
Define the arrangement of chromatograms
}
