% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{grid}
\alias{grid}
\title{Define a gridded plot.}
\usage{
grid(rows, cols, free_y = FALSE)
}
\arguments{
\item{rows}{The factors that represent rows.}

\item{cols}{The factors that represent columns.}

\item{free_y}{Whether the y-axis is free for each row.}
}
\value{
A function that takes and returns a lcmsPlotClass object.
}
\description{
Define a gridded plot.
}
\examples{
raw_files <- dir(
  system.file("cdf", package = "faahKO"),
  full.names = TRUE,
  recursive = TRUE
)[1:4]

metadata <- data.frame(
  sample_id = sub("\\\\.CDF", "", basename(raw_files)),
  factor1 = c("S", "S", "C", "C"),
  factor2 = c("T", "U", "T", "U")
)

p <- lcmsPlot(raw_files, metadata = metadata) +
  lcmsPlot::chromatogram(features = rbind(c(mzmin = 334.9, mzmax = 335.1, rtmin = 2700, rtmax = 2900))) +
  grid(rows = "factor1", cols = "factor2")
}
