% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-raw-data.R
\name{io_close_raw_data}
\alias{io_close_raw_data}
\title{Close a list of mzR object connections.}
\usage{
io_close_raw_data(raw_data)
}
\arguments{
\item{raw_data}{A list of mzR object connections.}
}
\value{
NULL.
}
\description{
Close a list of mzR object connections.
}
