% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\name{io_get_raw_data}
\alias{io_get_raw_data}
\title{Get \code{MsRawReader} objects for a set of sample paths}
\usage{
io_get_raw_data(sample_paths)
}
\arguments{
\item{sample_paths}{A \code{character} vector of file paths to the
raw MS data files (e.g., \code{.mzML}, \code{.mzXML}, \code{.CDF}, \code{.raw}).}
}
\value{
A named list of \code{MsRawReader} objects, where each element
corresponds to a file in \code{sample_paths}.
}
\description{
Get \code{MsRawReader} objects for a set of sample paths
}
\keyword{internal}
