% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-raw-data.R
\name{io_get_raw_data}
\alias{io_get_raw_data}
\title{Get the mzR objects for a set of sample paths.}
\usage{
io_get_raw_data(sample_paths)
}
\arguments{
\item{sample_paths}{The sample paths.}
}
\value{
A list of mzR objects.
}
\description{
Get the mzR objects for a set of sample paths.
}
