% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{is_xcms_object}
\alias{is_xcms_object}
\alias{is_xcms_object,lcmsPlotDataContainer-method}
\title{Check if the lcmsPlotDataContainer data object is an xcms object.}
\usage{
is_xcms_object(object)

\S4method{is_xcms_object}{lcmsPlotDataContainer}(object)
}
\arguments{
\item{object}{The lcmsPlotDataContainer object.}
}
\value{
Whether the data object is an xcms object.
}
\description{
Check if the lcmsPlotDataContainer data object is an xcms object.
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:5]

data_container <- create_data_container_from_obj(
  data_obj = raw_files,
  sample_id_column = NULL,
  metadata = NULL
)

is_xcms_object(data_container)
}
