% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{is_xcms_processed_data}
\alias{is_xcms_processed_data}
\title{Check whether an object is from XCMS and processed (e.g., peak detection).}
\usage{
is_xcms_processed_data(obj)
}
\arguments{
\item{obj}{The input object.}
}
\value{
Whether the object is an XCMS one.
}
\description{
Check whether an object is from XCMS and processed (e.g., peak detection).
}
