% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{iterate_plot_batches}
\alias{iterate_plot_batches}
\alias{iterate_plot_batches,lcmsPlotClass,function-method}
\title{Iterate on the batches of plots}
\usage{
iterate_plot_batches(object, iter_fn)

\S4method{iterate_plot_batches}{lcmsPlotClass,function}(object, iter_fn)
}
\arguments{
\item{object}{The lcmsPlotClass object}

\item{iter_fn}{The function to apply to each item being iterated on}
}
\description{
Iterate on the batches of plots
}
