% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakly-class.R
\name{labels}
\alias{labels}
\title{Define the labels of the plot, such as title and legend}
\usage{
labels(...)
}
\arguments{
\item{title}{The plot title}

\item{legend}{The legend's title}
}
\value{
A function that takes and returns a Peakly object
}
\description{
Define the labels of the plot, such as title and legend
}
