% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{layout}
\alias{layout}
\title{Define the plot layout}
\usage{
layout(design = NULL)
}
\arguments{
\item{design}{Specification of the location of areas in the layout (see https://patchwork.data-imaginist.com/reference/wrap_plots.html#arg-design)}
}
\description{
Define the plot layout
}
