% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{lcmsPlot}
\alias{lcmsPlot}
\title{Create a lcmsPlotClass object}
\usage{
lcmsPlot(dataset, sample_id_column = "sample_id")
}
\arguments{
\item{dataset}{An object of type XCMSnExp or MsExperiment}

\item{sample_id_column}{Which column should be used as the sample ID}
}
\value{
A lcmsPlotClass object
}
\description{
Create a lcmsPlotClass object
}
