% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\docType{class}
\name{lcmsPlotClass-class}
\alias{lcmsPlotClass-class}
\title{lcmsPlotClass class.}
\description{
lcmsPlotClass class.
}
\section{Slots}{

\describe{
\item{\code{options}}{The object options.}

\item{\code{data}}{The lcmsPlotDataContainer object.}

\item{\code{history}}{The list of the applied layers.}

\item{\code{plot}}{The underlying plot object.}
}}

