% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{mass_trace}
\alias{mass_trace}
\title{Define the mass trace to plot}
\usage{
mass_trace()
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define the mass trace to plot
}
