% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{merge_by_index}
\alias{merge_by_index}
\title{Merge two data frames by an index column.}
\usage{
merge_by_index(a, b, index_col)
}
\arguments{
\item{a}{The first data frame.}

\item{b}{The second data frame.}

\item{index_col}{The indexing column in data frame 'b'.}
}
\value{
The merged data frame.
}
\description{
Merge two data frames by an index column.
}
