% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\name{ms_header}
\alias{ms_header}
\alias{ms_header,MzrReader-method}
\alias{ms_header,RawrrReader-method}
\alias{ms_header,XcmsRawReader-method}
\title{Get scan header information from a raw MS file reader}
\usage{
ms_header(reader)

\S4method{ms_header}{MzrReader}(reader)

\S4method{ms_header}{RawrrReader}(reader)

\S4method{ms_header}{XcmsRawReader}(reader)
}
\arguments{
\item{reader}{An instance of \code{MsRawReader}.}
}
\value{
A \code{tibble} with at least the columns \code{seqNum},
\code{retentionTime}, \code{msLevel}, \code{basePeakIntensity}, and \code{totIonCurrent}.
}
\description{
Get scan header information from a raw MS file reader
}
\keyword{internal}
