% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\name{ms_peaks}
\alias{ms_peaks}
\alias{ms_peaks,MzrReader-method}
\alias{ms_peaks,RawrrReader-method}
\alias{ms_peaks,XcmsRawReader-method}
\title{Get peak matrices for selected scans from a raw MS file reader}
\usage{
ms_peaks(reader, scans)

\S4method{ms_peaks}{MzrReader}(reader, scans)

\S4method{ms_peaks}{RawrrReader}(reader, scans)

\S4method{ms_peaks}{XcmsRawReader}(reader, scans)
}
\arguments{
\item{reader}{An instance of \code{MsRawReader}.}

\item{scans}{An \code{integer} vector of scan sequence numbers to retrieve.}
}
\value{
A \code{list} of two-column matrices with columns \code{mz} and \code{intensity},
one element per scan (always a list, even for a single scan).
}
\description{
Get peak matrices for selected scans from a raw MS file reader
}
\keyword{internal}
