% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source-ms-dial.R
\name{MsDialPeaksSource}
\alias{MsDialPeaksSource}
\title{Create an MS-DIAL data source from peak lists}
\usage{
MsDialPeaksSource(peaks_paths, sample_paths, metadata_path = NULL)
}
\arguments{
\item{peaks_paths}{A \code{character} vector of paths to MS-DIAL
peak list files (CSV or TSV). The ordering should match the one
in the metadata (i.e., the samples).}

\item{sample_paths}{A \code{character} vector of paths to raw sample files
(e.g., \code{mzML}).}

\item{metadata_path}{A \code{character} value (optional) indicating the
path to a metadata file (CSV or TSV).
If \code{NULL}, a metadata \code{data.frame} is created from \code{sample_paths}.}
}
\value{
An object of class \code{ExternalDataSource}.
}
\description{
This function reads MS-DIAL peak list files and sample metadata
and converts them into a format compatible with xcms-style peak tables.
}
