% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source-mzmine.R
\name{MZmineFeatureListsSource}
\alias{MZmineFeatureListsSource}
\title{Create an MZmine data source from feature lists}
\usage{
MZmineFeatureListsSource(
  feature_lists_paths,
  sample_paths,
  metadata_path = NULL
)
}
\arguments{
\item{feature_lists_paths}{A \code{character} vector of paths to MZmine
feature list files (CSV or TSV).}

\item{sample_paths}{A \code{character} vector of paths to raw sample files
(e.g., \code{mzML}).}

\item{metadata_path}{A \code{character} value (optional) indicating the
path to a metadata file (CSV or TSV).
If \code{NULL}, a metadata \code{data.frame} is created from \code{sample_paths}.}
}
\value{
An object of class \code{ExternalDataSource}.
}
\description{
This function reads MZmine feature list files (supports version 2 and above)
and sample metadata and converts them into a format compatible
with xcms-style peak tables.
}
