% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{next_plot}
\alias{next_plot}
\alias{next_plot,lcmsPlotClass-method}
\title{Set the plot for an lcmsPlotClass object}
\usage{
next_plot(object)

\S4method{next_plot}{lcmsPlotClass}(object)
}
\arguments{
\item{object}{The lcmsPlotClass object}
}
\description{
Set the plot for an lcmsPlotClass object
}
