% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{next_plot}
\alias{next_plot}
\alias{next_plot,lcmsPlotClass-method}
\title{Set the plot for an lcmsPlotClass object}
\usage{
next_plot(object)

\S4method{next_plot}{lcmsPlotClass}(object)
}
\arguments{
\item{object}{The lcmsPlotClass object}
}
\description{
Set the plot for an lcmsPlotClass object
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:5]

p <- lcmsPlot(raw_files, batch_size = 2) +
  chromatogram(features = rbind(c(mzmin = 334.9, mzmax = 335.1, rtmin = 2700, rtmax = 2900))) +
  arrange(group_by = "sample_id") +
  legend(position = "bottom") +
  labels(legend = "Sample")

p <- next_plot(p)
}
