% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data}
\alias{plot_data}
\title{Plot the processed data (e.g., chromatograms)}
\usage{
plot_data(datasets, obj)
}
\arguments{
\item{datasets}{A list of data frames each corresponding to a dataset to plot.}

\item{obj}{The lcmsPlot object.}
}
\value{
The patchwork plot object.
}
\description{
Plot the processed data (e.g., chromatograms)
}
