% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-intensity-map.R
\name{plot_intensity_map}
\alias{plot_intensity_map}
\title{Plot a 2D intensity map.}
\usage{
plot_intensity_map(datasets, supporting_datasets, options, single = FALSE)
}
\arguments{
\item{datasets}{A list of data frames each corresponding to a dataset to plot.}

\item{supporting_datasets}{The supporting datasets (e.g., detected peaks).}

\item{options}{The plot object's options.}

\item{single}{Whether it is a single dataset variant.}
}
\value{
The plot as a ggplot2 object.
}
\description{
Plot a 2D intensity map.
}
