% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-variants.R
\name{plot_multiple_datasets}
\alias{plot_multiple_datasets}
\title{Plot multiple dataset types.}
\usage{
plot_multiple_datasets(datasets, obj, plot_config)
}
\arguments{
\item{datasets}{A list of data frames each corresponding to a dataset to plot.}

\item{obj}{The lcmsPlot object.}

\item{plot_config}{The plot configuration.
A list that specifies the plot function to use for each dataset type.}
}
\value{
The patchwork plot object.
}
\description{
Plot multiple dataset types.
}
