% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-variants.R
\name{plot_single_dataset}
\alias{plot_single_dataset}
\title{Plot a single dataset type.}
\usage{
plot_single_dataset(datasets, obj, plot_config)
}
\arguments{
\item{datasets}{A list of data frames each corresponding to a dataset to plot.}

\item{obj}{The lcmsPlot object.}

\item{plot_config}{The plot configuration.
A list that specifies the plot function to use for each dataset type.}
}
\value{
The patchwork plot object.
}
\description{
Plot a single dataset type.
}
