% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-total-ion-current.R
\name{plot_total_ion_current}
\alias{plot_total_ion_current}
\title{Plot total ion current.}
\usage{
plot_total_ion_current(datasets, supporting_datasets, options, single = FALSE)
}
\arguments{
\item{datasets}{A list of data frames each corresponding to a dataset to plot.}

\item{supporting_datasets}{The supporting datasets (e.g., detected peaks).}

\item{options}{The plot object's options.}

\item{single}{Whether it is a single dataset variant.}
}
\value{
The plot as a ggplot2 object.
}
\description{
Plot total ion current.
}
