% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{+,lcmsPlotClass,function-method}
\alias{+,lcmsPlotClass,function-method}
\title{Apply a function to the lcmsPlotClass object.}
\usage{
\S4method{+}{lcmsPlotClass,function}(e1, e2)
}
\arguments{
\item{e1}{A lcmsPlotClass object.}

\item{e2}{A function that takes a lcmsPlotClass object and returns another.}
}
\description{
Apply a function to the lcmsPlotClass object.
}
