% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source.R
\name{process_metadata}
\alias{process_metadata}
\title{Process sample metadata}
\usage{
process_metadata(sample_paths, metadata_path = NULL)
}
\arguments{
\item{sample_paths}{A \code{character} vector of sample file paths.}

\item{metadata_path}{A \code{character} value indicating the path to a delimited
metadata file with a header.}
}
\value{
A \code{data.frame} containing a \code{sample_path} column and
any additional metadata.
}
\description{
Construct a metadata data frame aligned with a set of sample paths. If a
metadata file is provided, it is read from disk and combined with the sample
paths. Otherwise, a minimal metadata table is created.
}
\keyword{internal}
