% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-file-reader.R
\docType{class}
\name{RawrrReader-class}
\alias{RawrrReader-class}
\title{A \code{rawrr}-backed raw MS file reader}
\description{
Stores the path to a ThermoFisher \code{.raw} file. \code{rawrr} does not maintain
persistent connections, so the path is re-used for each read operation.
}
\section{Slots}{

\describe{
\item{\code{path}}{A \code{character} value giving the path to the \code{.raw} file.}
}}

\keyword{internal}
