% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-data.R
\name{remove_null_elements}
\alias{remove_null_elements}
\title{Remove NULL values from a list.}
\usage{
remove_null_elements(lst)
}
\arguments{
\item{lst}{The input list.}
}
\value{
The list without NULL values.
}
\description{
Remove NULL values from a list.
}
