% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{rt_line}
\alias{rt_line}
\title{Define a vertical line across a retention time value}
\usage{
rt_line(intercept, line_type = "dashed", color = "black")
}
\arguments{
\item{intercept}{The x-axis intercept}

\item{line_type}{The line type}

\item{color}{The line color}
}
\description{
Define a vertical line across a retention time value
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:4]

p <- lcmsPlot(raw_files) +
  lcmsPlot::chromatogram(features = rbind(c(mzmin = 334.9, mzmax = 335.1, rtmin = 2700, rtmax = 2900))) +
  facets(facets = 'sample_id', ncol = 4) +
  rt_line(intercept = 2800, line_type = 'solid', color = 'red')
}
