% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{run_matching_plot_variant}
\alias{run_matching_plot_variant}
\title{Execute the suitable plot variant given the datasets and options.}
\usage{
run_matching_plot_variant(datasets, obj)
}
\arguments{
\item{datasets}{A list of data frames each corresponding to a dataset to plot.}

\item{obj}{The lcmsPlot object.}
}
\value{
The patchwork plot object.
}
\description{
Execute the suitable plot variant given the datasets and options.
}
