% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{show,lcmsPlotClass-method}
\alias{show,lcmsPlotClass-method}
\title{Plot the lcmsPlotClass object}
\usage{
\S4method{show}{lcmsPlotClass}(object)
}
\arguments{
\item{object}{The lcmsPlotClass object}
}
\description{
Plot the lcmsPlotClass object
}
