% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{show,lcmsPlotDataContainer-method}
\alias{show,lcmsPlotDataContainer-method}
\title{Show a summary of an instance of class \code{lcmsPlotDataContainer}}
\usage{
\S4method{show}{lcmsPlotDataContainer}(object)
}
\arguments{
\item{object}{An instance of class \code{lcmsPlotDataContainer}.}
}
\value{
Invisible \code{NULL}
}
\description{
Show a summary of an instance of class \code{lcmsPlotDataContainer}
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:5]

data_obj <- new("lcmsPlotDataContainer",
    data_obj = raw_files,
    metadata = data.frame(),
    chromatograms = data.frame(),
    mass_traces = data.frame(),
    spectra = data.frame(),
    total_ion_current = data.frame(),
    intensity_maps = data.frame(),
    rt_diff = data.frame(),
    feature_metadata = data.frame(),
    detected_peaks = data.frame())
data_obj
}
