% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlotDataContainer-class.R
\name{show,lcmsPlotDataContainer-method}
\alias{show,lcmsPlotDataContainer-method}
\title{Show a summary of an instance of class \code{lcmsPlotDataContainer}}
\usage{
\S4method{show}{lcmsPlotDataContainer}(object)
}
\arguments{
\item{object}{An instance of class \code{lcmsPlotDataContainer}.}
}
\value{
Invisible \code{NULL}
}
\description{
Show a summary of an instance of class \code{lcmsPlotDataContainer}
}
\examples{
raw_files <- dir(
   system.file("cdf", package = "faahKO"),
   full.names = TRUE,
   recursive = TRUE)[1:5]

data_obj <- new("lcmsPlotDataContainer",
    data_obj = raw_files,
    metadata = tibble::tibble(),
    chromatograms = tibble::tibble(),
    mass_traces = tibble::tibble(),
    spectra = tibble::tibble(),
    peak_density = tibble::tibble(),
    total_ion_current = tibble::tibble(),
    intensity_maps = tibble::tibble(),
    rt_diff = tibble::tibble(),
    feature_metadata = tibble::tibble(),
    detected_peaks = tibble::tibble())
data_obj
}
