% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{spectra}
\alias{spectra}
\title{Define the spectra to plot}
\usage{
spectra(
  sample_ids = NULL,
  mode = "closest_apex",
  ms_level = 1,
  rt = NULL,
  interval = 3,
  spectral_match_db = NULL,
  match_target_index = NULL
)
}
\arguments{
\item{sample_ids}{The sample IDs to consider. If NULL it will use the chromatogram ones.}

\item{mode}{The method to choose the scan. One of: closest, closest_apex, across_peak.}

\item{ms_level}{The MS level to consider for the scan.}

\item{rt}{The RT to consider - mode=closest}

\item{interval}{The RT interval to consider - mode=across_peak}

\item{spectral_match_db}{The spectral database to match against}

\item{match_target_index}{The target index for the mirror plot (index from the highest scoring)}
}
\value{
A function that takes and returns a lcmsPlotClass object
}
\description{
Define the spectra to plot
}
