% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{total_ion_current}
\alias{total_ion_current}
\title{Define the total ion currents.}
\usage{
total_ion_current(sample_ids = NULL, type = "boxplot")
}
\arguments{
\item{sample_ids}{The sample IDs to select.}

\item{type}{The type of plot; one of "boxplot", "violin", "jitter".}
}
\value{
A function that takes and returns a lcmsPlotClass object.
}
\description{
Define the total ion currents.
}
\examples{
data_obj <- get_XCMSnExp_object_example()

p <- lcmsPlot(data_obj, sample_id_column = "sample_name") +
  total_ion_current(type = "violin") +
  arrange(group_by = "sample_id") +
  labels(title = "Total ion current", legend = "Sample ID")
}
