% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsPlot-class.R
\name{total_ion_current}
\alias{total_ion_current}
\title{Define the total ion currents.}
\usage{
total_ion_current(sample_ids = NULL, type = "boxplot")
}
\arguments{
\item{sample_ids}{The sample IDs to select.}

\item{type}{The type of plot; one of "boxplot", "violin", "jitter".}
}
\description{
Define the total ion currents.
}
